+++ nn/initializer/initializer.go
@@ -1,7 +1,7 @@
 package initializer
 
 type Initializer interface {
-	Rand() float32
-	RandN(n int) []float32
-	RandShape(shapes ...int64) []float32
+	Rand() float64
+	RandN(n int) []float64
+	RandShape(shapes ...int64) []float64
 }
