+++ nn/initializer/normal.go
@@ -23,26 +23,26 @@ func NewNormal(mean, stddev float64) *Normal {
 	}
 }
 
-func (rand *Normal) Rand() float32 {
-	return float32(rand.n.Rand())
+func (rand *Normal) Rand() float64 {
+	return rand.n.Rand()
 }
 
-func (rand *Normal) RandN(n int) []float32 {
-	ret := make([]float32, n)
+func (rand *Normal) RandN(n int) []float64 {
+	ret := make([]float64, n)
 	for i := 0; i < n; i++ {
-		ret[i] = float32(rand.n.Rand())
+		ret[i] = rand.n.Rand()
 	}
 	return ret
 }
 
-func (rand *Normal) RandShape(shapes ...int64) []float32 {
+func (rand *Normal) RandShape(shapes ...int64) []float64 {
 	size := int64(1)
 	for _, s := range shapes {
 		size *= s
 	}
-	ret := make([]float32, size)
+	ret := make([]float64, size)
 	for i := 0; i < int(size); i++ {
-		ret[i] = float32(rand.n.Rand())
+		ret[i] = rand.n.Rand()
 	}
 	return ret
 }
